changed file and name space of v1p2 tests, integrated new oneroster_testcase to old test files
