--- classes/local/v1p2/tests/local/csv_client_test.php
@@ -1,38 +1,0 @@
-<?php
-// This file is part of Moodle - http://moodle.org/
-//
-// Moodle is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Moodle is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-
-/**
- * One Roster Enrolment Client Unit tests.
- *
- * @package    enrol_oneroster
- * @copyright  QUT Capstone Team - Abhinav Gandham, Harrison Dyba, Jonathon Foo, Kushi Patel
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-namespace enrol_oneroster;
-use enrol_oneroster\csv_client_test as csv_client_test_version_one;
-
-/**
- * One Roster tests for the client_helper class.
- *
- * @package    enrol_oneroster
- * @copyright  QUT Capstone Team - Abhinav Gandham, Harrison Dyba, Jonathon Foo, Kushi Patel
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- *
- * @covers  \enrol_oneroster\local\csv_client_helper
- */
-class csv_client_test extends csv_client_test_version_one {
-    // Insert new logic here.
-}
