+++ tests/local/v1p2/client_helper_test.php
@@ -22,21 +22,21 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace enrol_oneroster\local\v1p2;
+namespace enrol_oneroster\tests\local\v1p2;
 
-defined('MOODLE_INTERNAL') || die;
-require_once(__DIR__ . '/entity_testcase.php');
-use enrol_oneroster\local\entities\academic_session_test as academic_session_test_version_one;
+require_once('/var/www/moodle/enrol/oneroster/tests/client_helper_test.php');
+use enrol_oneroster\client_helper_test as client_helper_test_version_one;
+
 
 /**
- * One Roster tests for the academic_session entity.
+ * One Roster tests for the client_helper class.
  *
  * @package    enrol_oneroster
  * @copyright  QUT Capstone Team - Abhinav Gandham, Harrison Dyba, Jonathon Foo, Kushi Patel
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- * @covers  \enrol_oneroster\local\entity
- * @covers  \enrol_oneroster\local\entities\academic_session
+ *
+ * @covers  \enrol_oneroster\client_helper
  */
-class academic_session_test extends academic_session_test_version_one {
+class client_helper_test extends client_helper_test_version_one{
     // Insert new logic here.
 }
