+++ tests/local/v1p2/command_test.php
@@ -22,11 +22,11 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace enrol_oneroster\local\v1p2;
+namespace enrol_oneroster\tests\local\v1p2;
 
-defined('MOODLE_INTERNAL') || die;
-require_once(__DIR__ . '/oneroster_testcase.php');
-use enrol_oneroster\local\command_test as command_test_version_one;
+
+require_once('/var/www/moodle/enrol/oneroster/tests/local/command_test.php');
+use enrol_oneroster\tests\local\command_test as command_test_version_one;
 
 /**
  * One Roster tests for the `command` class.
