+++ tests/local/v1p2/csv_client_test.php
@@ -13,22 +13,26 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
 /**
  * One Roster Enrolment Client Unit tests.
  *
  * @package    enrol_oneroster
- * @copyright  QUT Capstone Team - Abhinav Gandham, Harrison Dyba,
- *             Jonathon Foo, Khushi Patel
+ * @copyright  QUT Capstone Team - Abhinav Gandham, Harrison Dyba, Jonathon Foo, Kushi Patel
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+namespace enrol_oneroster\tests\local\v1p2;
+use enrol_oneroster\tests\csv_test_helper as csv_client_test_version_one;
 
-
-namespace enrol_oneroster\local;
-
-defined('MOODLE_INTERNAL') || die();
-
-use advanced_testcase;
-
-class oneroster_testcase extends advanced_testcase {
-    //logic 
+/**
+ * One Roster tests for the client_helper class.
+ *
+ * @package    enrol_oneroster
+ * @copyright  QUT Capstone Team - Abhinav Gandham, Harrison Dyba, Jonathon Foo, Kushi Patel
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ *
+ * @covers  \enrol_oneroster\local\csv_client_helper
+ */
+class csv_client_test extends csv_client_test_version_one {
+    // Insert new logic here.
 }
