+++ tests/local/v1p2/entities/school_test.php
@@ -22,21 +22,22 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace enrol_oneroster\local\v1p2;
+namespace enrol_oneroster\tests\local\v1p2\entities;
 
-defined('MOODLE_INTERNAL') || die;
-require_once(__DIR__ . '/entity_testcase.php');
-use enrol_oneroster\local\entities\org_test as test_one;
+require_once(__DIR__ . '/../oneroster_testcase.php');
+use enrol_oneroster\tests\local\v1p2\oneroster_testcase;
+
+use enrol_oneroster\local\entities\school_test as school_test_version_one;
 
 /**
- * One Roster tests for the org entity.
+ * One Roster tests for the school entity.
  *
  * @package    enrol_oneroster
  * @copyright  QUT Capstone Team - Abhinav Gandham, Harrison Dyba, Jonathon Foo, Khushi Patel
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  * @covers  \enrol_oneroster\local\entity
- * @covers  \enrol_oneroster\local\entities\org
+ * @covers  \enrol_oneroster\local\entities\school
  */
-class org_test extends test_one {
+class school_test extends oneroster_testcase {
     // Insert new logic here.
 }
